[DP] A markdown explaining combined numeric, categorical, and TF-IDF features into final dataset, saved processed files and encoders, and added quality check with class distribution visualization.
A previous commit called "cleaning repo" was made because the git log was not aligned properly hence causing push and pull errors. the issue was resolved by Farai Rubvuta who had uploaded the last two cells containing the codes described on my behalf in haste knowing the deadline was looming on the horizon, forgive the disorganization in the repo commit logos.